requirements.txt
flask==3.0.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.0
open-clip-torch==2.20.0
pandas==2.1.0
scikit-learn==1.3.0